first test added!

+++ code/core/schema/schema_repository.py
@@ -1,5 +1,5 @@
-from code.core.schema.my_types import SchemaEnum
-from pyspark.sql.types import StructType, StructField, IntegerType, StringType, DateType, ArrayType
+from core.schema.my_types import SchemaEnum
+from pyspark.sql.types import StructType, StructField, LongType, StringType, DateType, ArrayType
 
 
 class SchemaRepository:
@@ -10,12 +10,12 @@ class SchemaRepository:
         match type:
             case SchemaEnum.ADDRESS:
                 schema = StructType([
-                    StructField("Unit", IntegerType(), True),
-                    StructField("Street No", IntegerType(), True),
+                    StructField("Unit", LongType(), True),
+                    StructField("Street No", LongType(), True),
                     StructField("Street Name", StringType(), True),
                     StructField("Suburb", StringType(), True),
                     StructField("State", StringType(), True),
-                    StructField("Postal Code", IntegerType(), True)
+                    StructField("Postal Code", LongType(), True)
                 ])
             case SchemaEnum.CITIZENSHIP:
                 inner_schema = StructType([
